api/backend/requirements.txt
Flask==2.3.2
Flask-Caching==2.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
scikit-learn==1.3.0
scipy==1.11.2
openai==0.27.9
numpy==1.25.2